# APIS/needle-haystack/openapi/python-flask/openapi_server/openapi/openapi.yaml
openapi: 3.0.2
info:
  license:
    name: MIT
  title: API Inspector
  version: 1.0.0
servers:
- url: http://localhost:8080/openapi-jaxrs-server-1.0.0/api/v3
paths:
...
  /mine:
    post:
      description: This can only be done by the logged in user.
      operationId: create_mine
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Mine'
          application/xml:
            schema:
              $ref: '#/components/schemas/Mine'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Mine'
        description: Created mine object
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mine'
            application/xml:
              schema:
                $ref: '#/components/schemas/Mine'
          description: successful operation
      summary: Create Mine
      tags:
      - mine
      x-openapi-router-controller: openapi_server.controllers.mine_controller
  /mine/createWithList:
    post:
      description: Creates list of mine with given input array
      operationId: create_mines_with_list_input
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: '#/components/schemas/Mine'
              type: array
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mine'
            application/xml:
              schema:
                $ref: '#/components/schemas/Mine'
          description: Successful operation
        default:
          description: successful operation
      summary: Creates list of Mine with given input array
      tags:
      - mine
      x-swagger-router-controller: ai.genpen.controllers.MineController
      x-openapi-router-controller: openapi_server.controllers.mine_controller
  /mine/{name}:
    delete:
      description: This can only be done by the logged in user.
      operationId: delete_mine
      parameters:
      - description: The mine that needs to be deleted by name
        explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "400":
          description: Invalid Mine variable supplied
        "404":
          description: Mine not found
      summary: Delete mine
      tags:
      - mine
      x-openapi-router-controller: openapi_server.controllers.mine_controller
    get:
      description: ""
      operationId: get_mine_by_name
      parameters:
      - description: 'The name that needs to be fetched. Use mine1 for testing. '
        explode: false
        in: path
        name: name
        required: true
        schema:
          format: string
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mine'
            application/xml:
              schema:
                $ref: '#/components/schemas/Mine'
          description: successful operation
        "400":
          description: Invalid Mine variable supplied
        "404":
          description: Mine not found
      summary: Get mine by name
      tags:
      - mine
      x-openapi-router-controller: openapi_server.controllers.mine_controller
    put:
      description: This can only be done by the logged in user.
      operationId: update_mine
      parameters:
      - description: name of mine that needs to be deleted
        explode: false
        in: path
        name: name
        required: true
        schema:
          format: string
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Mine'
          application/xml:
            schema:
              $ref: '#/components/schemas/Mine'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Mine'
        description: Update an existent mine in the system
      responses:
        default:
          description: successful operation
      summary: Update mine
      tags:
      - mine
      x-swagger-router-controller: ai.genpen.controllers.MineController
      x-openapi-router-controller: openapi_server.controllers.mine_controller
...
components:
  schemas:
...
    Mine:
      example:
        mined: false
        fk_sender_id: 6027456
        capacity_unit: capacity_unit
        length: 0
        active: false
        date_mined: 01-01-1968
        commodity_type: commodity_type
        capacity: 0
        risk_level: 0
        power_usage: 0.8008281904610115
        miner_id: 5962133
        width: 0
        name: "10"
        fk_owner_id: 1465812
        duration_its_been_active: 0
        created_date: 01-01-1968
        id: 119
      properties:
        capacity_unit:
          description: This attribute is a variable named capacity_unit
          format: string
          title: capacity_unit
          type: string
          example: null
        capacity:
          description: This attribute is a variable named capacity
          example: 0
          format: integer
          maximum: 10000000
          minimum: 0
          title: capacity
          type: integer
        power_usage:
          description: This attribute is a variable named power_usage
          format: integer
          title: power_usage
          type: number
        duration_its_been_active:
          description: This attribute is a variable named duration_its_been_active
          example: 0
          format: integer
          maximum: 10000000
          minimum: 0
          title: duration_its_been_active
          type: integer
        active:
          description: This attribute is a variable named active
          example: false
          format: boolean
          title: active
          type: boolean
        fk_sender_id:
          description: This attribute is a variable named fk_sender_id
          format: integer
          maximum: 10000000
          minimum: 0
          title: fk_sender_id
          type: integer
        risk_level:
          description: This attribute is a variable named risk_level
          example: 0
          format: integer
          title: risk_level
          type: integer
        created_date:
          description: This attribute is a variable named created_date
          example: 01-01-1968
          format: string
          title: created_date
          type: string
        fk_owner_id:
          description: This attribute is a variable named fk_owner_id
          format: integer
          maximum: 10000000
          minimum: 0
          title: fk_owner_id
          type: integer
        mined:
          description: This attribute is a variable named mined
          example: false
          format: boolean
          title: mined
          type: boolean
        length:
          description: This attribute is a variable named length
          example: 0
          format: integer
          maximum: 10000000
          minimum: 0
          title: length
          type: integer
        id:
          description: This is a GenPen.AI specific attribute
          example: 119
          format: integer
          maximum: 10000000
          minimum: 0
          title: id
          type: integer
        date_mined:
          description: This attribute is a variable named date_mined
          example: 01-01-1968
          format: string
          title: date_mined
          type: string
        width:
          description: This attribute is a variable named width
          example: 0
          format: integer
          maximum: 10000000
          minimum: 0
          title: width
          type: integer
        name:
          description: This is a GenPen.AI specific attribute
          example: "10"
          format: string
          title: name
          type: string
        commodity_type:
          description: This attribute is a variable named commodity_type
          format: string
          title: commodity_type
          type: string
          example: null
        miner_id:
          description: This attribute is a variable named miner_id
          format: integer
          maximum: 10000000
          minimum: 0
          title: miner_id
          type: integer
      title: Mine
      type: object
      xml:
        name: mine
      x-swagger-router-model: ai.genpen.models.Mine
...